Fix corner cases, refactor code for better readability
