+++ aidial_adapter_bedrock/llm/converse/adapter.py
@@ -1,24 +1,10 @@
-import json
-from typing import (
-    Any,
-    AsyncIterator,
-    Awaitable,
-    Callable,
-    Dict,
-    List,
-    Tuple,
-    cast,
-)
+from typing import Any, Awaitable, Callable, List, Tuple, cast
 
-from aidial_sdk.chat_completion import FinishReason as DialFinishReason
-from aidial_sdk.chat_completion import FunctionCall as DialFunctionCall
 from aidial_sdk.chat_completion import Message as DialMessage
-from aidial_sdk.chat_completion import ToolCall as DialToolCall
 
 from aidial_adapter_bedrock.bedrock import Bedrock
 from aidial_adapter_bedrock.dial_api.request import ModelParameters
 from aidial_adapter_bedrock.dial_api.storage import FileStorage
-from aidial_adapter_bedrock.dial_api.token_usage import TokenUsage
 from aidial_adapter_bedrock.llm.chat_model import (
     ChatCompletionAdapter,
     keep_last,
@@ -28,9 +14,11 @@ from aidial_adapter_bedrock.llm.consumer import Consumer
 from aidial_adapter_bedrock.llm.converse.input import (
     get_converse_system_prompt,
     process_messages,
-    to_converse_finish_reason,
     to_converse_tools,
-    to_dial_finish_reason,
+)
+from aidial_adapter_bedrock.llm.converse.output import (
+    process_non_streaming,
+    process_streaming,
 )
 from aidial_adapter_bedrock.llm.converse.types import (
     ConverseDeployment,
@@ -58,55 +46,6 @@ class ConverseAdapter(ChatCompletionAdapter):
         [ConverseDeployment, ConverseParams],
         Callable[[List[Tuple[ConverseMessage, Any]]], Awaitable[int]],
     ]
-
-    async def _process_streaming(
-        self, stream: AsyncIterator[Any], consumer: Consumer
-    ) -> None:
-        current_tool_use = None
-
-        async for event in stream:
-            if (content_block_start := event.get("contentBlockStart")) and (
-                tool_use := content_block_start.get("start", {}).get("toolUse")
-            ):
-                if current_tool_use is not None:
-                    raise ValueError("Tool use already started")
-                current_tool_use = {"input": ""} | tool_use
-
-            elif content_block := event.get("contentBlockDelta"):
-                delta = content_block.get("delta", {})
-
-                if message := delta.get("text"):
-                    consumer.append_content(message)
-
-                if "toolUse" in delta:
-                    if current_tool_use is None:
-                        raise ValueError(
-                            "Received tool delta before start block"
-                        )
-                    else:
-                        current_tool_use["input"] += delta["toolUse"].get(
-                            "input", ""
-                        )
-
-            elif event.get("contentBlockStop"):
-                if current_tool_use:
-                    consumer.create_function_tool_call(
-                        tool_call=DialToolCall(
-                            type="function",
-                            id=current_tool_use["toolUseId"],
-                            index=None,
-                            function=DialFunctionCall(
-                                name=current_tool_use["name"],
-                                arguments=current_tool_use["input"],
-                            ),
-                        )
-                    )
-                    current_tool_use = None
-
-            elif (message_stop := event.get("messageStop")) and (
-                stop_reason := message_stop.get("stopReason")
-            ):
-                consumer.close_content(to_dial_finish_reason(stop_reason))
 
     async def _discard_messages(
         self, params: ConverseParams, max_prompt_tokens: int | None
@@ -154,48 +93,19 @@ class ConverseAdapter(ChatCompletionAdapter):
         )
         return discarded_messages
 
-    def _process_non_streaming(
-        self, response: Dict[str, Any], consumer: Consumer
-    ) -> None:
-        message = response["output"]["message"]
-        for content_block in message.get("content", []):
-            if "text" in content_block:
-                consumer.append_content(content_block["text"])
-            if "toolUse" in content_block:
-                consumer.create_function_tool_call(
-                    tool_call=DialToolCall(
-                        type="function",
-                        id=content_block["toolUse"]["toolUseId"],
-                        index=None,
-                        function=DialFunctionCall(
-                            name=content_block["toolUse"]["name"],
-                            arguments=json.dumps(
-                                content_block["toolUse"]["input"]
-                            ),
-                        ),
-                    )
-                )
-
-        if usage := response.get("usage"):
-            consumer.add_usage(
-                TokenUsage(
-                    prompt_tokens=usage.get("inputTokens", 0),
-                    completion_tokens=usage.get("outputTokens", 0),
-                )
-            )
-
-        if stop_reason := response.get("stopReason"):
-            consumer.close_content(to_dial_finish_reason(stop_reason))
-
     async def construct_converse_params(
         self,
         messages: List[DialMessage],
         params: ModelParameters,
     ) -> ConverseParams:
         system_message = get_converse_system_prompt(messages)
+        processed_messages = await process_messages(messages, self.storage)
+        if not processed_messages.list:
+            raise ValidationError("List of messages must not be empty")
+
         return ConverseParams(
             system=[system_message] if system_message else None,
-            messages=await process_messages(messages, self.storage),
+            messages=processed_messages,
             inferenceConfig=cast(
                 InferenceConfig,
                 remove_nones(
@@ -203,12 +113,7 @@ class ConverseAdapter(ChatCompletionAdapter):
                         "temperature": params.temperature,
                         "topP": params.top_p,
                         "maxTokens": params.max_tokens,
-                        "stopSequences": [
-                            to_converse_finish_reason(
-                                DialFinishReason(finish_reason)
-                            ).value
-                            for finish_reason in params.stop
-                        ],
+                        "stopSequences": params.stop,
                     }
                 ),
             ),
@@ -239,7 +144,8 @@ class ConverseAdapter(ChatCompletionAdapter):
         consumer.set_discarded_messages(discarded_messages)
 
         if self.is_stream(params):
-            await self._process_streaming(
+            await process_streaming(
+                params=params,
                 stream=(
                     await self.bedrock.aconverse_streaming(
                         self.deployment, **converse_params.to_dict()
@@ -248,7 +154,8 @@ class ConverseAdapter(ChatCompletionAdapter):
                 consumer=consumer,
             )
         else:
-            self._process_non_streaming(
+            process_non_streaming(
+                params=params,
                 response=await self.bedrock.aconverse_non_streaming(
                     self.deployment, **converse_params.to_dict()
                 ),
